create new liboswkeys and move liboswlog to separate directory
