+++ lib/Makefile
@@ -1,12 +1,12 @@
 # Makefile for the KLIPS interface utilities
 # Copyright (C) 1998, 1999  Henry Spencer.
 # Copyright (C) 1999, 2000, 2001  Richard Guy Briggs
-# 
+#
 # This program is free software; you can redistribute it and/or modify it
 # under the terms of the GNU General Public License as published by the
 # Free Software Foundation; either version 2 of the License, or (at your
 # option) any later version.  See <http://www.fsf.org/copyleft/gpl.txt>.
-# 
+#
 # This program is distributed in the hope that it will be useful, but
 # WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
 # or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
@@ -17,7 +17,7 @@ include ${OPENSWANSRCDIR}/Makefile.inc
 
 srcdir?=$(shell pwd)/
 
-SUBDIRS=libopenswan libpluto libcrypto libwhack libipsecconf
+SUBDIRS=libopenswan libpluto libcrypto libwhack libipsecconf liboswlog liboswkeys
 
 ifeq ($(USE_LWRES),true)
 SUBDIRS+=libisc libdns liblwres
@@ -41,5 +41,5 @@ cleanall distclean mostlyclean realclean install programs checkprograms check cl
 	@for d in $(SUBDIRS) ; \
 	do \
 		(cd $$d && $(MAKE) srcdir=${srcdir}$$d/ OPENSWANSRCDIR=$(OPENSWANSRCDIR) $@ ) || exit 1;\
-	done; 
+	done;
