+++ lib/liboswkeys/oid.c
