+++ lib/liboswkeys/pgp.c
