+++ lib/liboswkeys/x509chain.c
