+++ lib/liboswlog/oswlog.c
