+++ programs/pluto/Makefile.options
@@ -54,7 +54,9 @@ endif
 
 # where to find klips headers and Openswan headers
 # and 2.6 kernel's <rtnetlink.h> and <xfrm.h>
-HDRDIRS = -I${OPENSWANSRCDIR}/programs/pluto/linux26 -I${OPENSWANSRCDIR}/include -I$(OPENSWANSRCDIR)/lib/libcrypto -I$(KLIPSINC)
+HDRDIRS = -I${OPENSWANSRCDIR}/programs/pluto/linux26
+HDRDIRS+= -I${OPENSWANSRCDIR}/include/pluto -I${OPENSWANSRCDIR}/include
+HDRDIRS+= -I$(OPENSWANSRCDIR)/lib/libcrypto -I$(KLIPSINC)
 
 # Use LIBNSS
 ifeq ($(USE_LIBNSS),true)
@@ -172,6 +174,9 @@ HAVE_THREADS_DIST_SRCS=${HAVE_THREADS_DIST_OBJS:.o=.c}
 X509_OBJS=${X509_DIST_OBJS}
 X509_SRCS=${X509_DIST_SRCS}
 X509_DEFS=-DX509_PLUTO
+X509_LIBS= ${LIBOSWKEYS} ${CRYPTOLIBS}
+LIBSWHACK+=${LIBOSWKEYS} ${CRYPTOLIBS}
+
 
 # dynamic LDAP CRL fetching requires OpenLDAP library
 ifeq ($(USE_LDAP),true)
@@ -301,7 +306,7 @@ DEFINES = $(EXTRA_DEFINES) \
 # Solaris 2 needs -lsocket -lnsl
 LIBSPLUTO =$(OBJSGCRYPT)
 LIBSPLUTO+=$(LIBOSWCRYPTO)
-LIBSPLUTO+=$(LIBDESLITE) $(LIBAES) $(OPENSWANLIB) $(LIBPLUTO)
+LIBSPLUTO+=$(LIBDESLITE) $(LIBAES) ${LIBOPENSWAN} ${LIBPLUTO}
 LIBSPLUTO+=${LIBSHA1} ${LIBMD5} ${LIBMD2}
 LIBSPLUTO+=$(IPSECPOLICY_LIBS) $(X509_LIBS)
 LIBSPLUTO+=$(HAVE_THREADS_LIBS)
@@ -311,6 +316,10 @@ LIBSPLUTO+=${WHACKLIB} ${PLUTOLIB} ${BSDKAME_LIBS}
 PLUTOMINUSL+= ${X509_LLIBS} ${CURL_LLIBS} ${TPM_LLIBS} ${HAVE_THREADS_LLIBS}
 PLUTOMINUSL+= ${XAUTH_LLIBS} ${XAUTHPAM_LIBS}
 PLUTOMINUSL+= ${OPENSSL_LIBS} ${LIBCRYPT} ${LIBGMP} # -lefence
+
+# Solaris needs -lsocket -lnsl
+LIBSWHACK+= ${WHACKLIB} ${LIBOPENSWAN}
+
 
 # Use LIBNSS
 ifeq ($(USE_LIBNSS),true)
