+++ programs/pluto/x509keys.c
@@ -57,100 +57,6 @@
 #include "pkcs.h"
 #include "x509more.h"
 
-/* extract id and public key from x.509 certificate and
- * insert it into a pubkeyrec
- */
-void
-add_x509_public_key(struct id *keyid
-		    , x509cert_t *cert
-		    , time_t until
-		    , enum dns_auth_level dns_auth_level)
-{
-    generalName_t *gn;
-    struct pubkey *pk;
-    cert_t c = { FALSE, CERT_X509_SIGNATURE, {cert} };
-
-    /* we support RSA only */
-    if (cert->subjectPublicKeyAlgorithm != PUBKEY_ALG_RSA) return;
-
-    /* ID type: ID_DER_ASN1_DN  (X.509 subject field) */
-    pk = allocate_RSA_public_key(c);
-    passert(pk != NULL);
-    pk->id.kind = ID_DER_ASN1_DN;
-    pk->id.name = cert->subject;
-    pk->dns_auth_level = dns_auth_level;
-    pk->until_time = until;
-    pk->issuer = cert->issuer;
-    delete_public_keys(&pluto_pubkeys, &pk->id, pk->alg);
-    install_public_key(pk, &pluto_pubkeys);
-
-    gn = cert->subjectAltName;
-
-    while (gn != NULL) /* insert all subjectAltNames */
-    {
-	struct id id = empty_id;
-
-	gntoid(&id, gn);
-	if (id.kind != ID_NONE)
-	{
-	    pk = allocate_RSA_public_key(c);
-	    pk->id = id;
-	    pk->dns_auth_level = dns_auth_level;
-	    pk->until_time = until;
-	    pk->issuer = cert->issuer;
-	    delete_public_keys(&pluto_pubkeys, &pk->id, pk->alg);
-	    install_public_key(pk, &pluto_pubkeys);
-	}
-	gn = gn->next;
-    }
-
-    if(keyid != NULL &&
-       keyid->kind != ID_DER_ASN1_DN &&
-       keyid->kind != ID_DER_ASN1_GN) {
-	pk = allocate_RSA_public_key(c);
-	pk->id = *keyid;
-
-	pk->dns_auth_level = dns_auth_level;
-	pk->until_time = until;
-	pk->issuer = cert->issuer;
-	delete_public_keys(&pluto_pubkeys, &pk->id, pk->alg);
-	install_public_key(pk, &pluto_pubkeys);
-    }
-}
-
-
-/*  when a X.509 certificate gets revoked, all instances of
- *  the corresponding public key must be removed
- */
-void
-remove_x509_public_key(/*const*/ x509cert_t *cert)
-{
-    const cert_t c = {FALSE, CERT_X509_SIGNATURE, {cert}};
-    struct pubkey_list *p, **pp;
-    struct pubkey *revoked_pk;
-
-    revoked_pk = allocate_RSA_public_key(c);
-    p          = pluto_pubkeys;
-    pp         = &pluto_pubkeys;
-
-    while(p != NULL)
-   {
-	if (same_RSA_public_key(&p->key->u.rsa, &revoked_pk->u.rsa))
-	{
-	    /* remove p from list and free memory */
-	    *pp = free_public_keyentry(p);
-	    loglog(RC_LOG_SERIOUS,
-		"invalid RSA public key deleted");
-	}
-	else
-	{
-	    pp = &p->next;
-	}
-	p =*pp;
-    }
-    free_public_key(revoked_pk);
-}
-
 /*
  * Decode the CERT payload of Phase 1.
  */
